Add `ReceiptOrderable`
Extract `receipt_order` update functionality from `DraftContentItem` to
enable reuse in `LiveContentItem`
